Do not check `hasPermissions()` when requesting permissions
It seems there's a bug in the Google Fit or the same behavior as on iOS (where `hasPermissions()` has no use). People reported this   for example here https://github.com/android/fit-samples/issues/28

Because of this the only solution at the moment is to not use the `hasPermissions()` at all. Unfortunately the Android implementation of the `requestAuthorization()` checks permissions the same way like for the `hasPermissions()` and doesn't allow to authorize the app again for Google Fit use because `hasPermissions()` returns `true` even if there are no rights at all. 

I've decided to post a PR to the Android implementation and just not check the permissions in the `requestAuthorization()`. From what I tested if permissions are assigned the autorization dialog is not shown at all so the permission check is in the `GoogleSignIn.requestPermissions()` itself and we don't need to check this again before calling it.